requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
contourpy==1.3.1
    # via matplotlib
customtkinter==5.2.2
    # via -r requirements.in
cycler==0.12.1
    # via matplotlib
darkdetect==0.8.0
    # via customtkinter
fonttools==4.55.3
    # via matplotlib
kiwisolver==1.4.8
    # via matplotlib
matplotlib==3.10.0
    # via -r requirements.in
mutagen==1.47.0
    # via -r requirements.in
mysql-connector-python==8.1.0
    # via -r requirements.in
numpy==2.0.0
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
packaging==24.2
    # via
    #   customtkinter
    #   matplotlib
pillow==11.1.0
    # via
    #   -r requirements.in
    #   matplotlib
protobuf==4.21.12
    # via mysql-connector-python
psutil==6.1.1
    # via -r requirements.in
pydub==0.25.1
    # via -r requirements.in
pygame==2.6.1
    # via -r requirements.in
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
python-dotenv==1.0.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil

# The following packages are considered to be unsafe in a requirements file:
# setuptools
